flask/email_authentication/requirements.txt
alembic==1.7.7
bcrypt==4.0.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
dominate==2.7.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.45
typing_extensions==4.1.1
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
